Deploy assissted installer using podman kube
This is to deploy the v2.X assisted installer versions which
include the iso service container.

Along with this additional vars for specifing hosts in assisted installer config
giving the user more control over the assited installer config. This allows
for configurations where the bastion and host need to talk to the assisted installer
via different IP addresses. By default they are both the same howerver you can now
set assisted_service_host to control the ip address which the nodes use to talk to
assisted installer.